+++ aio_overpass/element.py
@@ -2,13 +2,13 @@
 
 import math
 import re
-from abc import ABC, abstractmethod
 from collections import defaultdict
 from collections.abc import Iterable, Iterator
 from dataclasses import dataclass
 from typing import Any, Generic, TypeAlias, TypeVar, cast
 
 from aio_overpass import Query
+from aio_overpass.spatial import GeoJsonDict, Spatial
 
 import shapely.geometry
 import shapely.ops
@@ -19,28 +19,16 @@ from shapely.geometry.base import BaseGeometry, BaseMultipartGeometry
 __docformat__ = "google"
 __all__ = (
     "collect_elements",
-    "Spatial",
     "Element",
     "Node",
     "Way",
     "Relation",
     "Relationship",
+    "Bbox",
     "GeometryDetails",
     "Metadata",
-    "Bbox",
-    "GeoJsonDict",
-    "SpatialDict",
 )
 
-GeoJsonDict: TypeAlias = dict[str, Any]
-"""
-A dictionary representing a GeoJSON object.
-"""
-
-_OverpassDict: TypeAlias = dict[str, Any]
-"""
-A dictionary representing a JSON object returned by the Overpass API.
-"""
 
 Bbox: TypeAlias = tuple[float, float, float, float]
 """
@@ -51,65 +39,6 @@ This tuple can be understood as any of
     - ``(minlat, minlon, maxlat, maxlon)``
     - ``(minx, miny, maxx, maxy)``
 """
-
-
-@dataclass(kw_only=True, slots=True)
-class SpatialDict:
-    """
-    Mapping of spatial objects with the ``__geo_interface__`` property.
-
-    Objects of this class have the ``__geo_interface__`` property following a protocol
-    [proposed](https://gist.github.com/sgillies/2217756) by Sean Gillies, which can make
-    it easier to use spatial data in other Python software. An example of this is the ``shape()``
-    function that builds Shapely geometries from any object with the ``__geo_interface__`` property.
-
-    Attributes:
-        __geo_interface__: this is the proposed property that contains the spatial data
-    """
-
-    __geo_interface__: dict
-
-
-class Spatial(ABC):
-    """
-    Base class for (groups of) geospatial objects.
-
-    Classes that represent spatial features extend this class and implement the
-    ``geojson`` property. Exporting objects in the GeoJSON format should make it possible
-    to integrate them with other tools for visualization, or further analysis.
-    The ability to re-import the exported GeoJSON structures as ``Spatial`` objects is not
-    considered here.
-    """
-
-    __slots__ = ("__validated__",)  # we use that field in tests
-
-    @property
-    @abstractmethod
-    def geojson(self) -> GeoJsonDict:
-        """
-        A mapping of this object, using the GeoJSON format.
-
-        The coordinate reference system for all GeoJSON coordinates is ``CRS:84``,
-        which means every coordinate is a tuple of longitude and latitude (in that order)
-        on the WGS 84 ellipsoid. Note that this order is flipped for all Shapely geometries
-        that represent OpenStreetMap elements (latitude first, then longitude).
-
-        References:
-            - https://osmdata.openstreetmap.de/info/projections.html
-            - https://tools.ietf.org/html/rfc7946#section-4
-        """
-        raise NotImplementedError
-
-    @property
-    def geo_interfaces(self) -> Iterator[SpatialDict]:
-        """A mapping of this object to ``SpatialDict``s that implement ``__geo_interface__``."""
-        geojson = self.geojson
-        match geojson["type"]:
-            case "FeatureCollection":
-                for feature in geojson["features"]:
-                    yield SpatialDict(__geo_interface__=feature)
-            case _:
-                yield SpatialDict(__geo_interface__=geojson)
 
 
 @dataclass(kw_only=True, slots=True)
@@ -491,6 +420,9 @@ _ElementKey: TypeAlias = tuple[str, int]
 
 _MemberKey: TypeAlias = tuple[_ElementKey, str]
 """Relation members are identified by their element key and role."""
+
+_OverpassDict: TypeAlias = dict[str, Any]
+"""A dictionary representing a JSON object returned by the Overpass API."""
 
 
 class _ElementCollector: